more protocol byte tests
