+++ src/kafka-tests/kafka-tests.csproj
@@ -91,6 +91,7 @@
     <Compile Include="Unit\DefaultPartitionSelectorTests.cs" />
     <Compile Include="Unit\ProducerTests.cs" />
     <Compile Include="Unit\ProtocolAssertionExtensions.cs" />
+    <Compile Include="Unit\ProtocolByteTests.cs" />
     <Compile Include="Unit\ProtocolTests.cs" />
     <Compile Include="Unit\ProtocolBaseRequestTests.cs" />
     <Compile Include="Unit\ProtocolMessageTests.cs" />
